feitas conexões com banco de dados, criadas DTOs e separado em camadas
