+++ modulo6/projeto-labook/src/data/BaseDataBase.ts
@@ -1,0 +1,22 @@
+import Knex from "knex"
+import dotenv from "dotenv"
+
+dotenv.config()
+
+export abstract class BaseDataBase {
+
+    protected static connection: Knex = Knex({
+        client: "mysql",
+        connection: {
+           host: process.env.DB_HOST,
+           user: process.env.DB_USER,
+           password: process.env.DB_PASSWORD,
+           database: process.env.DB_SCHEMA,
+           port: 3306,
+           multipleStatements: true
+        }
+     })
+
+}
+
+
