+++ modulo6/projeto-labook/src/model/IdDTO.ts
@@ -1,0 +1,3 @@
+export interface authenticationData {
+    id: string
+ };
