+++ modulo6/projeto-labook/src/model/PostDTO.ts
@@ -1,0 +1,13 @@
+export enum POST_TYPES {
+    NORMAL = "normal",
+    EVENT = "event"
+ };
+ 
+export interface post {
+    postId: string,
+    photo: string,
+    description: string,
+    type: POST_TYPES,
+    createdAt: Date,
+    authorId: string
+ };
