+++ modulo6/projeto-labook/src/services/GenerateId.ts
@@ -1,0 +1,5 @@
+import { v4 } from "uuid"
+
+export const generateId = (): string => {
+    return v4()
+}
